requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pypdf
langchain-google-genai
pinecone-client
langchain-pinecone
langchain-community
huggingface_hub
langchain sentence_transformers